fix: ci
cargo

smol ci fixes

fix: ci

fix: ci

fix: dprint

fix: udeps

fix: unit tests

clippy

lock
