+++ examples/bsc-sdk/src/node/cli.rs
@@ -1,0 +1,148 @@
+use crate::{
+    chainspec::{parser::BscChainSpecParser, BscChainSpec},
+    node::{consensus::BscConsensus, BscNode},
+};
+use clap::{value_parser, Parser};
+use reth::{
+    args::LogArgs,
+    builder::{NodeBuilder, WithLaunchContext},
+    cli::Commands,
+    prometheus_exporter::install_prometheus_recorder,
+    version::{LONG_VERSION, SHORT_VERSION},
+    CliRunner,
+};
+use reth_chainspec::EthChainSpec;
+use reth_cli::chainspec::ChainSpecParser;
+use reth_cli_commands::node::NoArgs;
+use reth_db::DatabaseEnv;
+
+use reth_network::EthNetworkPrimitives;
+use reth_node_ethereum::BasicBlockExecutorProvider;
+use reth_tracing::FileWorkerGuard;
+use std::{
+    fmt::{self},
+    future::Future,
+    sync::Arc,
+};
+use tracing::info;
+
+/// The main reth_bsc cli interface.
+///
+/// This is the entrypoint to the executable.
+#[derive(Debug, Parser)]
+#[command(author, version = SHORT_VERSION, long_version = LONG_VERSION, about = "Reth", long_about = None)]
+pub struct Cli<Spec: ChainSpecParser = BscChainSpecParser, Ext: clap::Args + fmt::Debug = NoArgs> {
+    /// The command to run
+    #[command(subcommand)]
+    pub command: Commands<Spec, Ext>,
+
+    /// The chain this node is running.
+    ///
+    /// Possible values are either a built-in chain or the path to a chain specification file.
+    #[arg(
+        long,
+        value_name = "CHAIN_OR_PATH",
+        long_help = Spec::help_message(),
+        default_value = Spec::SUPPORTED_CHAINS[0],
+        value_parser = Spec::parser(),
+        global = true,
+    )]
+    chain: Arc<Spec::ChainSpec>,
+
+    /// Add a new instance of a node.
+    ///
+    /// Configures the ports of the node to avoid conflicts with the defaults.
+    /// This is useful for running multiple nodes on the same machine.
+    ///
+    /// Max number of instances is 200. It is chosen in a way so that it's not possible to have
+    /// port numbers that conflict with each other.
+    ///
+    /// Changes to the following port numbers:
+    /// - `DISCOVERY_PORT`: default + `instance` - 1
+    /// - `AUTH_PORT`: default + `instance` * 100 - 100
+    /// - `HTTP_RPC_PORT`: default - `instance` + 1
+    /// - `WS_RPC_PORT`: default + `instance` * 2 - 2
+    #[arg(long, value_name = "INSTANCE", global = true, default_value_t = 1, value_parser = value_parser!(u16).range(..=200))]
+    instance: u16,
+
+    #[command(flatten)]
+    logs: LogArgs,
+}
+
+impl<C, Ext> Cli<C, Ext>
+where
+    C: ChainSpecParser<ChainSpec = BscChainSpec>,
+    Ext: clap::Args + fmt::Debug,
+{
+    /// Execute the configured cli command.
+    ///
+    /// This accepts a closure that is used to launch the node via the
+    /// [`NodeCommand`](reth_cli_commands::node::NodeCommand).
+    pub fn run<L, Fut>(self, launcher: L) -> eyre::Result<()>
+    where
+        L: FnOnce(WithLaunchContext<NodeBuilder<Arc<DatabaseEnv>, C::ChainSpec>>, Ext) -> Fut,
+        Fut: Future<Output = eyre::Result<()>>,
+    {
+        self.with_runner(CliRunner::try_default_runtime()?, launcher)
+    }
+
+    /// Execute the configured cli command with the provided [`CliRunner`].
+    pub fn with_runner<L, Fut>(mut self, runner: CliRunner, launcher: L) -> eyre::Result<()>
+    where
+        L: FnOnce(WithLaunchContext<NodeBuilder<Arc<DatabaseEnv>, C::ChainSpec>>, Ext) -> Fut,
+        Fut: Future<Output = eyre::Result<()>>,
+    {
+        // add network name to logs dir
+        self.logs.log_file_directory =
+            self.logs.log_file_directory.join(self.chain.chain().to_string());
+
+        let _guard = self.init_tracing()?;
+        info!(target: "reth::cli", "Initialized tracing, debug log directory: {}", self.logs.log_file_directory);
+
+        // Install the prometheus recorder to be sure to record all metrics
+        let _ = install_prometheus_recorder();
+
+        let _components = |spec: Arc<C::ChainSpec>| {
+            (BasicBlockExecutorProvider::new(spec.clone()), BscConsensus::new(spec))
+        };
+
+        match self.command {
+            Commands::Node(command) => {
+                runner.run_command_until_exit(|ctx| command.execute(ctx, launcher))
+            }
+            Commands::Init(command) => {
+                runner.run_blocking_until_ctrl_c(command.execute::<BscNode>())
+            }
+            Commands::InitState(command) => {
+                runner.run_blocking_until_ctrl_c(command.execute::<BscNode>())
+            }
+            Commands::DumpGenesis(command) => runner.run_blocking_until_ctrl_c(command.execute()),
+            Commands::Db(command) => runner.run_blocking_until_ctrl_c(command.execute::<BscNode>()),
+            Commands::Stage(_command) => todo!(),
+            Commands::P2P(command) => {
+                runner.run_until_ctrl_c(command.execute::<EthNetworkPrimitives>())
+            }
+            Commands::Config(command) => runner.run_until_ctrl_c(command.execute()),
+            Commands::Recover(command) => {
+                runner.run_command_until_exit(|ctx| command.execute::<BscNode>(ctx))
+            }
+            Commands::Prune(command) => runner.run_until_ctrl_c(command.execute::<BscNode>()),
+            Commands::Import(_command) => {
+                // runner.run_blocking_until_ctrl_c(command.execute::<BscNode, _, _>(components))
+                todo!()
+            }
+            Commands::Debug(_command) => todo!(),
+            #[cfg(feature = "dev")]
+            Commands::TestVectors(_command) => todo!(),
+        }
+    }
+
+    /// Initializes tracing with the configured options.
+    ///
+    /// If file logging is enabled, this function returns a guard that must be kept alive to ensure
+    /// that all logs are flushed to disk.
+    pub fn init_tracing(&self) -> eyre::Result<Option<FileWorkerGuard>> {
+        let guard = self.logs.init_tracing()?;
+        Ok(guard)
+    }
+}
